Created LoadEventsController and adjusted the rest of code
