+++ src/database/DatabaseDAO.java
@@ -12,7 +12,7 @@ import Entities.Temporary_entites.*;
 
 
 
-public class DatabaseDAO implements LoadMapDataAccessInterface {
+public class DatabaseDAO implements LoadEventsDataAccessInterface {
     Database database = new Database();
 
     public void addEvent(String event_name, String event_description, String type, String time, String date, String creator, String longitude, String latitude) {
