+++ src/interface_adapter/LoadEventsController.java
@@ -1,0 +1,50 @@
+package interface_adapter;
+
+import org.jxmapviewer.JXMapViewer;
+import org.jxmapviewer.viewer.GeoPosition;
+
+import java.awt.*;
+import java.awt.geom.Point2D;
+import java.util.HashSet;
+import java.util.Set;
+
+public class LoadEventsController {
+    JXMapViewer mapViewer;
+
+    public LoadEventsController(JXMapViewer mapViewer) {
+        this.mapViewer = mapViewer;
+    }
+
+    public boolean checkForClickOnEvent(Point clickPoint) {
+
+        //TODO: events are received from UseCaseInteractor
+        // In the line below instead of this here Set<Events> from UseCaseInteractor should be used
+
+        Set<Event> events = new HashSet<>();
+
+        // Check if any waypoints are clicked
+        for (Event event : events) {
+            GeoPosition eventPosition = event.getPosition();
+            GeoPosition clickedPosition = mapViewer.convertPointToGeoPosition(clickPoint);
+
+            // Check if the click is within a certain distance of the waypoint
+            if (isClickNearWaypoint(clickedPosition, eventPosition, mapViewer)) {
+                //TODO: switch to another UseCase (VIEW_EVENT)
+                return true;
+            }
+        }
+        return false;
+    }
+
+    private boolean isClickNearWaypoint(GeoPosition clickPosition, GeoPosition waypointPosition, JXMapViewer mapViewer) {
+        // Convert GeoPositions to screen coordinates
+        Point2D.Double clickPoint = (Point2D.Double) mapViewer.getTileFactory().geoToPixel(clickPosition, mapViewer.getZoom());
+        Point2D.Double waypointPoint = (Point2D.Double) mapViewer.getTileFactory().geoToPixel(waypointPosition, mapViewer.getZoom());
+
+        // Define a threshold for considering the click as near the waypoint
+        int threshold = 23;
+
+        // Check if the click is within the threshold distance of the waypoint
+        return clickPoint.distance(waypointPoint) < threshold;
+    }
+}
